Implemented Basic Menu Prototype
Implemented a basic menu prototype, with title text, a play button, and a quit button. clicking the play button changes the game to a different scene, currently set to a placeholder but can be changed with a variable. clicking the quit button prints a message to the unity console and disables all content in the current scene.